Escuela Nivel_Modalidad y Escuela_Otras_Ofertas, y Ciclo_Lectivo

+++ database/migrations/2023_12_28_230411_create_ciclo_lectivo_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('ciclo_lectivo', function (Blueprint $table) {
-            $table->smallInteger('id')->unsigned()->primary();
+            $table->tinyInteger('id')->unsigned()->primary();
             $table->year('nombre');
             $table->tinyInteger('orden');
             $table->boolean('vigente');
